+++ scripts/tests/pipeline/test_validate_pipeline_env.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from pipeline.multi_env_runner import resolve_env_selection
+from pipeline.pipeline_parameters import resolve_env_names
 
 
 @pytest.fixture(autouse=True)
@@ -17,7 +17,7 @@ class TestResolveEnvSelection:
     @pytest.mark.unit
     def test_fails_when_no_env_selection(self):
         with pytest.raises(ValueError, match="Set ENV_NAMES or both CLUSTER_NAME and ENVIRONMENT_NAME"):
-            resolve_env_selection()
+            resolve_env_names()
 
     @pytest.mark.unit
     def test_fails_when_env_names_combined_with_cluster_name(self, monkeypatch):
@@ -25,7 +25,7 @@ class TestResolveEnvSelection:
         monkeypatch.setenv("CLUSTER_NAME", "cluster-01")
 
         with pytest.raises(ValueError, match="Set ENV_NAMES only, or both CLUSTER_NAME and ENVIRONMENT_NAME, but not both at the same time"):
-            resolve_env_selection()
+            resolve_env_names()
 
     @pytest.mark.unit
     def test_fails_when_env_names_combined_with_environment_name(self, monkeypatch):
@@ -33,7 +33,7 @@ class TestResolveEnvSelection:
         monkeypatch.setenv("ENVIRONMENT_NAME", "env-01")
 
         with pytest.raises(ValueError, match="Set ENV_NAMES only, or both CLUSTER_NAME and ENVIRONMENT_NAME, but not both at the same time"):
-            resolve_env_selection()
+            resolve_env_names()
 
     @pytest.mark.unit
     def test_fails_when_env_names_combined_with_both_per_env_vars(self, monkeypatch):
@@ -42,13 +42,13 @@ class TestResolveEnvSelection:
         monkeypatch.setenv("ENVIRONMENT_NAME", "env-01")
 
         with pytest.raises(ValueError, match="Set ENV_NAMES only, or both CLUSTER_NAME and ENVIRONMENT_NAME, but not both at the same time"):
-            resolve_env_selection()
+            resolve_env_names()
 
     @pytest.mark.unit
     def test_passes_with_env_names_only(self, monkeypatch):
         monkeypatch.setenv("ENV_NAMES", "cluster-01/env-01")
 
-        result = resolve_env_selection()
+        result = resolve_env_names()
 
         assert result == ["cluster-01/env-01"]
         assert os.environ["ENV_NAMES"] == "cluster-01/env-01"
@@ -58,14 +58,14 @@ class TestResolveEnvSelection:
         monkeypatch.setenv("ENV_NAMES", "cluster-01-env-01")
 
         with pytest.raises(ValueError, match="Invalid environment name"):
-            resolve_env_selection()
+            resolve_env_names()
 
     @pytest.mark.unit
     def test_synthesises_env_names_from_cluster_and_environment_name(self, monkeypatch):
         monkeypatch.setenv("CLUSTER_NAME", "cluster-01")
         monkeypatch.setenv("ENVIRONMENT_NAME", "env-01")
 
-        result = resolve_env_selection()
+        result = resolve_env_names()
 
         assert result == ["cluster-01/env-01"]
         assert os.environ["ENV_NAMES"] == "cluster-01/env-01"
@@ -75,4 +75,4 @@ class TestResolveEnvSelection:
         monkeypatch.setenv("CLUSTER_NAME", "cluster-01")
 
         with pytest.raises(ValueError, match="Set ENV_NAMES or both CLUSTER_NAME and ENVIRONMENT_NAME"):
-            resolve_env_selection()
+            resolve_env_names()
